+++ changelog.rst
@@ -1,7 +1,7 @@
 Changelog
 =========
 
-Last change: 06-NOV-2025 CEST
+Last change: 18-NOV-2025 CEST
 
 0.9.0
 -----
@@ -12,13 +12,14 @@ Important updates:
 - Removed dependency ``h5py``. Picasso now uses ``pandas`` to read and write HDF5 files
 - Updated other dependencies, most importantly, ``numpy`` is now in version 2
 - Old setup files were replaced by ``pyproject.toml`` for building and packaging Picasso
-- Black-based code formatting applied to all scripts
+- New option to save cluster areas/volumes in DBSCAN, HDBSCAN and SMLM clusterer
 - Render: unfold groups was removed as it is contained within the square grid unfolding
 - SPINNA: improved saved fit results summary (see issue #560)
 
 *Small improvements and bug fixes:*
 +++++++++++++++++++++++++++++++++++
 
+- Black-based code formatting applied to all scripts
 - Render: unfold groups/picks (rectangular grid) fixed for nonconsecutive grouping (the grid might have had missing elements before)
 - Render: apply drift from external file fixed
 - Render: fix masking (issue #560)